+++ packages/mutation-testing-metrics-scala/circe/src/main/scala/mutationtesting/MutationReportEncoder.scala
@@ -2,28 +2,28 @@ package mutationtesting
 
 import io.circe._
 
-@deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
+@deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
 object MutationReportEncoder {
 
-  @deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
-  implicit lazy val mutantStatusEncoder: Encoder[MutantStatus] = MutationReportCodec.mutantStatusCodec
+  @deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
+  implicit lazy val mutantStatusEncoder: Encoder[MutantStatus] = circe.mutantStatusCodec
 
-  @deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
-  implicit lazy val positionEncoder: Encoder[Position] = MutationReportCodec.positionCodec
+  @deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
+  implicit lazy val positionEncoder: Encoder[Position] = circe.positionCodec
 
-  @deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
-  implicit lazy val thresholdsEncoder: Encoder[Thresholds] = MutationReportCodec.thresholdsCodec
+  @deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
+  implicit lazy val thresholdsEncoder: Encoder[Thresholds] = circe.thresholdsCodec
 
-  @deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
-  implicit lazy val locationEncoder: Encoder[Location] = MutationReportCodec.locationCodec
+  @deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
+  implicit lazy val locationEncoder: Encoder[Location] = circe.locationCodec
 
-  @deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
-  implicit lazy val mutantResultEncoder: Encoder[MutantResult] = MutationReportCodec.mutantResultCodec
+  @deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
+  implicit lazy val mutantResultEncoder: Encoder[MutantResult] = circe.mutantResultCodec
 
-  @deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
-  implicit lazy val fileResultEncoder: Encoder[FileResult] = MutationReportCodec.fileResultCodec
+  @deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
+  implicit lazy val fileResultEncoder: Encoder[FileResult] = circe.fileResultCodec
 
-  @deprecated(message = "Use MutationReportCodec instead", since = "1.5.0")
-  implicit lazy val mutationTestResultEncoder: Encoder[MutationTestResult] = MutationReportCodec.mutationTestResultCodec
+  @deprecated(message = "Use mutationtesting.circe instead", since = "1.5.0")
+  implicit lazy val mutationTestResultEncoder: Encoder[MutationTestResult] = circe.mutationTestResultCodec
     .mapJson(_.deepDropNullValues)
 }
